requirements.txt
opencv-python>=4.8.0
ultralytics>=8.0.0
anthropic>=0.18.0
pillow>=10.0.0
numpy>=1.24.0
matplotlib>=3.7.0
python-dotenv>=1.0.0
